Remove trailing newline on notebook output for PyCharm compatibility
`nbformat.write` ensures files always end in a newline (see:
[`nbformat` commit `a273247`: finish notebook files with a trailing newline](https://github.com/TeoZosa/nbformat/commit/a2732477e31a92888662ad7448d3e5ec5ac60a93)
), which seems to be the prevailing norm throughout the Jupyter
ecosystem; see:
- [`nbconvert`: Don't remove the newline when converting notebooks to notebooks #213](https://github.com/jupyter/nbconvert/pull/213)
- JupyterHub GUI: creating/saving/copying notebooks (tested)

On the other hand, PyCharm forcefully strips newlines from notebooks on
save operations (though interestingly ONLY on save operations and not,
for example, on notebook creation). This behavior seems to have been
implemented in
[PY-40629: PyCharm should save notebooks to the same JSON format Jupyter does](https://youtrack.jetbrains.com/issue/PY-40629)
to resolve
[PY-36262: Remove newline at the end of ipynb file and stem cell metadata in case it's empty](https://youtrack.jetbrains.com/issue/PY-36262)

Note: Given the above, the most appropriate course of action seems to be
 raising the issue with the PyCharm team directly as opposed to either
the `nbstripout` or `nbformat` teams.